backend/requirements.txt
fastapi
uvicorn
loguru
python-dotenv
pyyaml
requests
numpy
scipy
watchdog
